src: ensure that remove_master is not called with a grab active
We add a rather convoluted way to queue the master removal on the main
thread instead of doing it in the signal handler where the menu's
pointer grab is still active, messing up internal XServer state.

Closes #1

Another way of doing it would be to not use a menu at all but rather
some button or 'trash' icon directly in the tree view.